memory-service/requirements.txt
fastapi
uvicorn
mem0ai
qdrant-client
